Add a page parameter, that will dictate the attachments of which page are being displayed in the attachment picker #454
* moved ViewFileDisplay in ViewFile parameters class
